Kafka/client/requirements.txt
kafka-python==1.4.6
confluent_kafka
